Collisions were implemented on PhysicsEngine::processMovement.
PhysicsEngine now stores GameMap* for collisions(which are calculated only if GameMap* was installed using PhysicsEngine::initializeCollisionSystem).

PhysicsEngine::processAcceleration was removed. Its functionalyti is now part of PhysicsEngine::processMovement.
PhysicsEngine::processQueue was removed. Its functionality is now part of PhysicsEngine::loop.

compile_time macroses time_correction and speed_test were added to PhysicProcessingFunctions.cpp in order to make code easier to read.

Error about time-independent acceration and speeding was resolved. Now it correctly relies on PhysicsEngine::UpdateInterval.

Constants::g is now equal to 9.8f, as it was intended. Incorrect time conversion bug was resolved.

GameObjects::Vector is now called vector. It's using 'scalar' as its member type. scalar is currently equal to float, but it can be changed in the future.

Opertors *=, +, - and * were implemented for vector.

A set of unnecessary member functions was removed from AbstractGameObject. List of them:
- mass() const;
- accelerate(float, float);
- updateSpeed();
-updatePosition();

size_h and size_v were added to constructor of AbstractGameObject. These arguments are necessary to be passed from the creator of the object.

AbstractGameObject.cpp was added in order to store operators for ObjectState.

vector size was added to ObjectState. It stores dimentions of space ocupied by the object.

GameMap is now capable of returning the multiplier and renderInfo of the block, which is stored in the cell with given coordinates.

Structure of main() was changed. Now it loads settings and then gives control to specially created function game_process(Settings&) which initializes game elements and links them together.
It was done because main function was becoming too messy.